+++ lib/fake_consul.rb
@@ -1,5 +1,6 @@
 require "fake_consul/version"
-require "fake_consul/server"
+require "fake_consul/kv"
+require "fake_consul/service"
 
 module FakeConsul