+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v1.0.0] - 2021-01-22
+- rename Server to Kv
+- Add Service mocking
+
 ## [v0.1.0] - 2020-11-19
 - Update required gems
-- Returns just string instead of hash with `.get`
+- Returns a string instead of hash with `.get`
 
 ## [v0.0.7] - 2015-12-07
 - Adds `delete` command